do not cache scroll container (#46)
if there is enough space initially to show all tree elements,
the scroll container cannot be identified and a dummy is put
into the cache instead. When the screen space is reduced or the
tree grows (e.g. by expanding composite nodes) scrolling becomes
necessary, but won't work properly with the dummy container.